Updated location

+++ src/FETCH/FETCH_no_logs.ps1
@@ -280,7 +280,11 @@ function Add-WmiClassInstance {
     # Set properties dynamically
     foreach ($key in $Properties.Keys) {
         if ($instance.PSObject.Properties.Name -contains $key) {
-            $instance.$key = $Properties[$key]
+            $instance.$key = if ($Properties[$key] -match '^\d{4}-\d{2}-\d{2} \d{2}:\d{2} UTC$') {
+                ([DateTime]::ParseExact($Properties[$key], "yyyy-MM-dd HH:mm 'UTC'", [System.Globalization.CultureInfo]::InvariantCulture)).ToString("yyyyMMddHHmmss.ffffff+000")
+            } else {
+                $Properties[$key]
+            }
         } else {
             Write-Log "WARNING" "Property '$key' is not defined in the WMI class '$wmiClassName'. Skipping."
         }
@@ -322,7 +326,7 @@ function Remove-OldInstances {
         Write-Log "VERBOSE" "Cleanup complete. Remaining instances: $((Get-WmiObject -Namespace $WmiNamespace -Class $wmiClassName).Count)"
     }
     else {
-        Write-Log "VERBOSE" "Found no instances older than $DeleteOlderThanDays days. No cleanup needed."
+        Write-Log "VERBOSE" "Found no instances older than $DeleteOlderThanDays days in $WmiClassName. No cleanup needed."
     }
 }
 
@@ -487,7 +491,7 @@ try {
         $keyProp = @{ "CollectionDatetime" = "datetime" }
         $props = @{ 
             "UserSID" = "string"
-            "LastSeen" = "string"
+            "LastSeen" = "datetime"
             "ComputerSID" = "string"
         }
         Add-WmiClass -WmiNamespace $WmiNamespace -WmiClassPrefix $WmiClassPrefix -CollectionType "Sessions" -KeyProperty $keyProp -Properties $props
@@ -847,7 +851,7 @@ try {
                         Add-WmiClassInstance -WmiNamespace $WmiNamespace -WmiClassPrefix $WmiClassPrefix -CollectionType 'LocalGroups' -Properties $localGroupMember
                     }
                 } else {
-                    Write-Log "WARNING" "Skipping $memberType $memberSID in $($currentGroup.Name) ($($currentGroup.ObjectIdentifier))"
+                    Write-Log "VERBOSE" "Skipping $memberType $memberSID in $($currentGroup.Name) ($($currentGroup.ObjectIdentifier))"
                 }
             }
             # Add each local group to script output
@@ -956,7 +960,7 @@ try {
         $jsonOutput | Out-File $WriteTo
     }   
 
-    if ($DeleteOlderThanDays -ne $null) {
+    if ($Wmi -and $DeleteOlderThanDays -ne $null) {
         Remove-OldInstances -DeleteOlderThanDays $DeleteOlderThanDays -WmiNamespace $WmiNamespace -WmiClassName "$WmiClassPrefix`Sessions"
         Remove-OldInstances -DeleteOlderThanDays $DeleteOlderThanDays -WmiNamespace $WmiNamespace -WmiClassName "$WmiClassPrefix`UserRights"
         Remove-OldInstances -DeleteOlderThanDays $DeleteOlderThanDays -WmiNamespace $WmiNamespace -WmiClassName "$WmiClassPrefix`LocalGroups"
@@ -974,4 +978,4 @@ try {
         Set-PSDebug -Off
         Stop-Transcript
     }
-}
+}
